# clubApi-swagger/api/swagger/swagger.production.yaml
swagger: '2.0'
info:
  description: |
    This is the documentation of PacLab Backend for User Manager 
    [http://swagger.io](http://swagger.io) or on 
    [irc.freenode.net, #swagger](http://swagger.io/irc/). 
    Created by @pdonaire1.
  version: 1.0.0
  title: PacLab UserManager API
  termsOfService: http://swagger.io/terms/
  contact:
    email: pablo.donaire@azumo.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: user.clusters.paclabnw.com
basePath: /
schemes:
  - https
paths:
  /users/login:
    post:
      description: User Login
      operationId: UserLogin
      parameters:
        - name: email
          type: string
          in: query
          description: Email of the user to login
          required: true
        - name: password
          type: string
          in: query
          description: Password of the user to login
          required: true
      responses:
        200:
          description: Return the token, user object, status and message
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              user:
                $ref: '#/definitions/User'
              token:
                type: "string"
              message:
                type: "string"
        400:
          description: Unauthorized
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
  /users:
    put:
      security:
        - Bearer: []
      description: Update Current User
      operationId: UserUpdate
      parameters: 
        - name: Authorization
          type: string
          in: header
          description: User token provided
          required: true
        - name: firstName
          in: query
          description: User First Name
          type: string
        - name: lastName
          in: query
          description: User Last Name
          type: string
        - name: phone
          in: query
          description: User last name
          type: string
          required: false
      responses:
        200:
          description: Return User Updated status
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              user:
                $ref: "#/definitions/User"
              message:
                type: "string"
        400:
          description: Invalid input
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
              message:
                type: "string"
  /users/current:
    get:
      security:
        - Bearer: []
      description: Get info of the current logged user
      operationId: UserCurrent
      parameters: 
        - name: Authorization
          type: string
          in: header
          description: User token provided
          required: true
      responses:
        200:
          description: Return User Updated status
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              user:
                $ref: "#/definitions/User"
              message:
                type: "string"
        400:
          description: Invalid input
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
              message:
                type: "string"
  /users/change-password:
    post:
      security:
        - Bearer: []
      description: User can change the current password with this endpoint (authentication required)
      operationId: UserChangePass
      parameters:
        - name: Authorization
          type: string
          in: header
          description: User token provided
          required: true
        - name: currentPassword
          in: query
          description: User Current Password
          type: string
          required: true
        - name: newPassword
          in: query
          description: New User Password
          type: string
          required: true
      responses:
        200:
          description: Return the user, token and message when user change the password successfully
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              user:
                $ref: '#/definitions/User'
              token:
                type: "string"
              message:
                type: "string"
        400:
          description: Invalid input
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
              message:
                type: "string"
  /users/retrieve-password-step-one:
    post:
      description: If User forgot the password he can retrieve the password with this endpoint, this will send an email to the user to set a new password
      operationId: UserRetrievePass
      parameters:
        - name: email
          in: query
          description: User Current Password
          type: string
          required: true
      responses:
        200:
          description: Return info if user was fount and send an email to retrieve the password
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              message:
                type: "string"
        400:
          description: Invalid input
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
              message:
                type: "string"
  /users/retrieve-password-step-two:
    post:
      description: This endpoint is usefull, we can use this as the second part of the last step, or we can use is to set the password by first time, when an Admin creates a new User to the customer.
      operationId: UserRetrievePassTwo
      parameters:
        - name: token
          in: query
          description: User Token generated by the email
          type: string
          required: true
        - name: newPassword
          in: query
          description: User New password to be created
          type: string
          required: true
      responses:
        200:
          description: Shows user, token and status if the form is valid
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              message:
                type: "string"
              user:
                $ref: '#/definitions/User'
              token:
                type: "string"
        400:
          description: Invalid input
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
              message:
                type: "string"
  /users/create:
    post:
      security:
        - Bearer: []
      description: As Admin of Customer or as System Admin. Create User for customers, this will send an email to the user for create the new password. IMPORTANT NOTE the endpoint for the user created to set the password for the first time is the same than the step two of retrieve password (/users/retrieve-password-step-two)
      operationId: UserCreate
      parameters: 
        - name: Authorization
          type: string
          in: header
          description: User token provided
          required: true
        - name: email
          in: query
          description: User email
          type: string
          required: true
        - name: firstName
          in: query
          description: User first name 
          type: string
          required: true
        - name: lastName
          in: query
          description: User last name
          type: string
          required: true
        - name: phone
          in: query
          description: User last name
          type: string
          required: false
        - name: customerId
          in: query
          description: Id of the customer to be user associated (not required)
          type: string
        - name: roles
          in: query
          description: (Only if is System Admin), can pass the param role of the allowed roles ['SYS_ADMIN', 'ADMIN', 'USER']
          type: string
      responses:
        200:
          description: Return info if user is created successfully
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              token:
                type: "string"
              user:
                $ref: '#/definitions/User'
              
        400:
          description: Invalid input
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              error:
                type: "object"
              message:
                type: "string"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    type: "object"
    properties:
      _id:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      active:
        type: "boolean"
      company_account:
        type: "string"
      roles:
        type: "array"
        items:
          type: "string"
      createdAt:
        type: "string"
      updatedAt:
        type: "string"
    xml:
      name: "User"
...